initial refactor completed. now files need to be moved to proper folders
