+++ src/PlayPads.js
@@ -3,7 +3,7 @@ import PlayPad from "./PlayPad";
 import usePlayPadColors from "./usePlayPadColors";
 import { sortPadColorMap } from "./helperFunctions";
 import "./PlayPads.css";
-import { presetBankData } from "./PresetBank0";
+import { presetBankData } from "./PresetDesign";
 
 function PlayPads() {
   const { currentTrack, recState, recDispatch, padsRecMode, setPadsRecMode } = useContext(presetBankData);
