--- src/PlayerButton.css
@@ -1,18 +1,0 @@
-.donut__pads__padInactive{
-    margin: 2px;
-    background-color: grey;
-    padding: 4px;
-    height: 4em;
-    width:4em;
-    padding: 4px;
-    border: 2px solid black;
-}
-
-.donut__pads__padActive{
-    background-color: rgba(155, 47, 160, 0.728);
-    height: 4em;
-    width:4em;
-    border: 2px solid black;
-    padding: 4px;
-
-}
