--- src/PlayerPad.css
@@ -1,17 +1,0 @@
-.PlayerPad{
-    height: 10em;
-    width: 10em;
-    margin: 5px;
-    border-radius: 10%;
-}
-
-
-@media (max-width: 850px) {
-    .PlayerPad{
-        height: 6.75em;
-        width: 6.75em;
-        margin: 5px;
-        border-radius: 10%;
-
-    }
-  }
