--- src/Tester.css
@@ -1,141 +1,0 @@
-
-
-.practice{
-    height: 100vh;
-    background-color: #E5E5E5;
-}
-
-.transportButtons{
-    /* border: 2px solid green; */
-    display: flex;
-    justify-content: space-evenly;
-    padding-top: 1.5vh;
-    padding-bottom: 1vh;
-}
-
-.transportButtons__buttonBox{
-    display: flex;
-    flex-flow: column;
-    text-align: center;
-
-}
-
-#buttdescription{
-    margin-top: 4px;
-}
-
-.playerPads{
-    padding-top: 1vh;
-    display: flex;
-    justify-content: center;
-    margin: 1vh;
-    height: 50vh;
-}
-
-#transportIcon{
-    padding-top: 2.75px;
-}
-
-#transport__snapCircleContents{
-    display: flex;
-    padding-top: 9.75px;
-}
-
-
-
-.toggleSounds{
-    margin-top: 19vh;
-    border-top: 3.5px solid rgba(94, 59, 190, 0.295);
-    /* padding-bottom: 9vh; */
-}
-
-.toggleMode{
-    margin-top: 7vh;
-    border-top: 3.5px solid rgba(94, 59, 190, 0.295);
-    padding-bottom: 9vh;
-}
-
-.transportButtons__buttonBox__circleInactive{
-    height: 40px;
-    width: 40px;
-    border-radius: 50%;
-    border: 2px solid rgb(55, 0, 255);
-
-}
-
-.transportButtons__buttonBox__circleActive{
-    height: 40px;
-    width: 40px;
-    border-radius: 50%;
-    border: 4px solid rgb(0, 225, 255);
-
-}
-#buttonBox__description{
-    margin-top: 4px;
-    color: rgb(43, 16, 139);
-}
-
-@media (max-width: 850px) {
-    .playerPads{
-        padding-top: 5vh;
-        display: flex;
-        justify-content: center;
-        margin: 1vh;
-        height: 50vh;
-    }
-
-    .progressBar{
-        padding-top: 2vh;
-    }
-    .transportButtons{
-        /* border: 2px solid green; */
-        display: flex;
-        justify-content: space-evenly;
-        padding-top: 3vh;
-        padding-bottom: 1vh;
-    }
-
-
-    .toggleSounds{
-        margin-top: 5vh;
-        border-top: 3.5px solid rgba(94, 59, 190, 0.295);
-        /* padding-bottom: 9vh; */
-    }
-
-    .toggleMode{
-        margin-top: 10vh;
-        border-top: 3.5px solid rgba(94, 59, 190, 0.295);
-        padding-bottom: 9vh;
-    }
-
-    .transportButtons__buttonBox__circleInactive{
-        height: 40px;
-        width: 40px;
-        border-radius: 50%;
-        border: 2px solid rgb(55, 0, 255);
-    
-    }
-    
-    .transportButtons__buttonBox__circleActive{
-        height: 40px;
-        width: 40px;
-        border-radius: 50%;
-        border: 4px solid rgb(0, 225, 255);
-    
-    }
-
-    .transportButtons__buttonBox{
-        display: flex;
-        flex-flow: column;
-        text-align: center;
-    }
-
-    #buttonBox__description{
-        margin-top: 4px;
-        color: rgb(43, 16, 139);
-
-    }
-
-
-
-  }
