--- src/useActiveArray.js
@@ -1,10 +1,0 @@
-import React, { useState } from "react";
-
-function useActiveArray() {
-  const [padStates, setPadStates] = useState([]);
-
-
-  return [padColors, padColorsContructor];
-}
-
-export default useActiveArray;
